更新(upload.sh)：不再备份整个项目，只备份rpm包

See merge request lixuebing/cve-fixer!9

+++ upload.sh
@@ -80,8 +80,7 @@ git push -uf origin $branch
 cd ..
 
 # 做个备份
-rm -rf success/Package/$package_name
-# 创建目录
-mkdir success/{Package,RPMS} -p
-cp -rf $package_name/RPMS success/
-cp -rf $package_name success/Package/
+if [ -f $package_name/RPMS ]; then
+	mkdir success
+	cp -rf $package_name/RPMS success/
+fi
